Relocate `BuiltFromProposedBlock` to `utils.rs`
The `BuiltFromProposedBlock` definitions is going to be referenced in `GlobalState`.
In order to prepare for this change the struct needs to be moved to a different
crate in order to avoid circular dependencies.
